requirements.txt
colorama==0.4.6
psycopg2==2.9.7
python-dotenv==1.0.0
